Move fetching of templates/authorizations to handlers directory
