+++ packages/node/src/core/evm/handlers/index.ts
@@ -1,2 +1,3 @@
+export * from './fetch-templates-authorizations';
 export * from './initialize-provider';
 export * from './process-transactions';
